app/models/ai/audio_analyzer.py
"""
Audio Analyzer Module

This module provides audio analysis capabilities for ScholarIQ, including
transcription, speaker diarization, and audio feature extraction.
"""

import logging
import io
import wave
import numpy as np
from typing import Any, Dict, List, Optional, Tuple, Union

# Conditional imports for audio processing
try:
    import librosa
    import soundfile as sf
    AUDIO_DEPS_AVAILABLE = True
except ImportError:
    AUDIO_DEPS_AVAILABLE = False

from .base_analyzer import BaseAnalyzer

logger = logging.getLogger(__name__)

class AudioAnalyzer(BaseAnalyzer):
    """
    Audio analysis implementation for ScholarIQ.
    
    This analyzer processes audio content to extract transcripts, speaker information,
    and other audio features relevant for educational content analysis.
    """
    
    def __init__(self, model_name: Optional[str] = None, **kwargs):
        """
        Initialize the audio analyzer.
        
        Args:
            model_name: Name of the model to use for speech recognition.
            **kwargs: Additional model configuration parameters.
        """
        if not AUDIO_DEPS_AVAILABLE:
            raise ImportError(
                "Audio processing dependencies not found. "
                "Please install with: pip install librosa soundfile"
            )
            
        super().__init__(model_name=model_name, **kwargs)
        self.sample_rate = 16000  # Default sample rate in Hz
        self.max_duration = 600  # Maximum audio duration to process (10 minutes)
    
    def _load_model(self, **kwargs) -> Any:
        """
        Load the audio processing model.
        
        In a production environment, this would load a pre-trained model
        for speech recognition and audio analysis.
        
        Args:
            **kwargs: Additional model configuration parameters.
            
        Returns:
            The loaded model or None if using direct processing.
        """
        # In a real implementation, this would load a pre-trained model
        # For example: return whisper.load_model(self.model_name or "base")
        return None
    
    async def analyze(
        self,
        audio_data: Union[bytes, str],
        **kwargs
    ) -> Dict[str, Any]:
        """
        Analyze the provided audio content.
        
        Args:
            audio_data: Audio data as bytes or file path.
            **kwargs: Additional analysis parameters.
                - language: Expected language of the audio (e.g., 'en', 'es')
                - diarize: Whether to perform speaker diarization (default: False)
                - sentiment: Whether to analyze sentiment (default: False)
                
        Returns:
            A dictionary containing the analysis results.
        """
        try:
            # Load and preprocess audio
            audio_array, sample_rate = await self._load_audio(audio_data)
            duration = len(audio_array) / sample_rate
            
            # Limit processing duration
            if duration > self.max_duration:
                logger.warning(f"Audio duration ({duration:.1f}s) exceeds maximum allowed duration. Truncating.")
                max_samples = int(self.max_duration * sample_rate)
                audio_array = audio_array[:max_samples]
                duration = self.max_duration
            
            # Basic audio features
            features = {
                "duration_seconds": duration,
                "sample_rate": sample_rate,
                "channels": 1,  # Assuming mono
                "rms_energy": float(np.sqrt(np.mean(audio_array**2))),
                "zero_crossing_rate": float(np.mean(librosa.feature.zero_crossing_rate(audio_array))),
            }
            
            # Perform speech recognition if requested
            if kwargs.get("transcribe", True):
                try:
                    transcript = await self._transcribe_audio(audio_array, sample_rate, **kwargs)
                    features["transcript"] = transcript
                except Exception as e:
                    logger.error(f"Speech recognition failed: {str(e)}")
                    features["transcript"] = None
            
            # Speaker diarization if requested
            if kwargs.get("diarize", False):
                try:
                    diarization = await self._diarize_speakers(audio_array, sample_rate, **kwargs)
                    features["speakers"] = diarization
                except Exception as e:
                    logger.error(f"Speaker diarization failed: {str(e)}")
            
            # Sentiment analysis if requested
            if kwargs.get("sentiment", False) and "transcript" in features and features["transcript"]:
                try:
                    sentiment = await self._analyze_sentiment(features["transcript"])
                    features["sentiment"] = sentiment
                except Exception as e:
                    logger.error(f"Sentiment analysis failed: {str(e)}")
            
            return features
            
        except Exception as e:
            logger.exception("Error in audio analysis")
            raise ValueError(f"Audio analysis failed: {str(e)}")
    
    async def _load_audio(
        self,
        audio_data: Union[bytes, str],
        target_sample_rate: int = 16000
    ) -> Tuple[np.ndarray, int]:
        """
        Load and preprocess audio data.
        
        Args:
            audio_data: Audio data as bytes or file path.
            target_sample_rate: Target sample rate for the audio.
            
        Returns:
            Tuple of (audio_array, sample_rate)
        """
        try:
            if isinstance(audio_data, bytes):
                # Load from bytes
                with io.BytesIO(audio_data) as audio_file:
                    audio_array, sample_rate = sf.read(audio_file, dtype='float32')
            else:
                # Load from file path
                audio_array, sample_rate = sf.read(audio_data, dtype='float32')
            
            # Convert stereo to mono if needed
            if len(audio_array.shape) > 1:
                audio_array = np.mean(audio_array, axis=1)
            
            # Resample if needed
            if sample_rate != target_sample_rate:
                audio_array = librosa.resample(
                    audio_array,
                    orig_sr=sample_rate,
                    target_sr=target_sample_rate
                )
                sample_rate = target_sample_rate
            
            # Normalize audio
            audio_array = librosa.util.normalize(audio_array)
            
            return audio_array, sample_rate
            
        except Exception as e:
            logger.error(f"Error loading audio: {str(e)}")
            raise ValueError(f"Could not load audio: {str(e)}")
    
    async def _transcribe_audio(
        self,
        audio_array: np.ndarray,
        sample_rate: int,
        **kwargs
    ) -> str:
        """
        Transcribe audio to text.
        
        Args:
            audio_array: Audio samples as a numpy array.
            sample_rate: Sample rate of the audio.
            **kwargs: Additional parameters for transcription.
                - language: Expected language code (e.g., 'en', 'es')
                
        Returns:
            Transcribed text.
        """
        # In a production environment, this would use a speech recognition model
        # For example: return self.model.transcribe(audio_array, language=kwargs.get('language', 'en'))
        
        # This is a placeholder implementation
        logger.warning("Speech recognition not implemented. Returning placeholder transcript.")
        return "[Audio transcription would appear here in a production environment]"
    
    async def _diarize_speakers(
        self,
        audio_array: np.ndarray,
        sample_rate: int,
        **kwargs
    ) -> List[Dict[str, Any]]:
        """
        Perform speaker diarization on the audio.
        
        Args:
            audio_array: Audio samples as a numpy array.
            sample_rate: Sample rate of the audio.
            **kwargs: Additional parameters for diarization.
                - min_speakers: Minimum number of speakers to detect
                - max_speakers: Maximum number of speakers to detect
                
        Returns:
            List of speaker segments with start/end times and speaker IDs.
        """
        # In a production environment, this would use a speaker diarization model
        # For example: return diarize(audio_array, sample_rate, **kwargs)
        
        # This is a placeholder implementation
        logger.warning("Speaker diarization not implemented. Returning placeholder data.")
        return [
            {
                "start": 0.0,
                "end": len(audio_array) / sample_rate,
                "speaker": "SPEAKER_00",
                "confidence": 0.95
            }
        ]
    
    async def _analyze_sentiment(self, text: str) -> Dict[str, float]:
        """
        Analyze sentiment of transcribed text.
        
        Args:
            text: Text to analyze.
            
        Returns:
            Sentiment analysis results.
        """
        # In a production environment, this would use a sentiment analysis model
        # For example: return sentiment_model.analyze(text)
        
        # This is a placeholder implementation
        return {
            "positive": 0.5,
            "negative": 0.1,
            "neutral": 0.4,
            "compound": 0.4
        }
    
    def get_capabilities(self) -> Dict[str, Any]:
        """Get information about the audio analyzer's capabilities."""
        capabilities = super().get_capabilities()
        capabilities.update({
            "content_types": [
                "audio/wav",
                "audio/mp3",
                "audio/mpeg",
                "audio/ogg",
                "audio/webm"
            ],
            "features": [
                "transcription",
                "speaker_diarization",
                "audio_features",
                "sentiment_analysis"
            ],
            "dependencies_available": AUDIO_DEPS_AVAILABLE,
            "max_duration_seconds": self.max_duration
        })
        return capabilities
