requirements.txt
# Core Dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
aiofiles>=23.2.1

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary==2.9.10

# AI/ML
numpy>=1.24.0
pandas>=2.1.0
Pillow>=10.0.0
torch>=2.1.0
transformers[sentencepiece]>=4.35.0
opencv-python>=4.8.1
librosa>=0.10.1
spacy>=3.7.2

# Deployment
gunicorn==21.2.0

# Testing
pytest==7.4.3
httpx==0.25.1
pytest-cov==4.1.0

# Code Quality
black==23.11.0
isort==5.12.0
mypy==1.7.0
dnspython==2.8.0
email-validator==2.3.0